gadakeco-neat/requirements.txt
pygame>=1.9.6
pillow>=6.1.0
